services/api/requirements.txt
fastapi==0.104.1
mangum==0.17.0
boto3==1.34.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
